Disable desktop icons for a short period of time #111
When openbox changes the mouse cursor to a progress pointer,
via the sm_notify linux library, a Mouse grab is requested
to X11 server.

This momentarily disables the desktop icons, also
sets the hourglass cursor when hovering over the icons,
which was not happening until now - its was confusingly
set to the regular pointer cursor.